requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
azure-cognitiveservices-speech==1.32.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
config==0.5.1
contourpy==1.1.1
cycler==0.12.0
DateTime==5.2
decorator==5.1.1
elevenlabs==0.2.26
executing==1.2.0
filelock==3.12.4
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.35
gTTS==2.3.2
idna==3.4
importlib-metadata==6.8.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mpyg321==2.1.1
msgpack==1.0.7
multidict==6.0.4
networkx==3.1
noisereduce==3.0.0
np==1.0.2
numba==0.58.0
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
pocketsphinx==5.0.2
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
scikit-learn==1.3.1
scipy==1.11.3
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.0
stack-data==0.6.2
streamlit==1.27.2
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
Wave==0.0.2
wcwidth==0.2.6
webrtcvad==2.0.10
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
zope.interface==6.1
